Move `UniqueShapeContext1960` to feature types header
