+++ common/src/feature_types.cpp
@@ -76,4 +76,14 @@ namespace pcl
       os << (i == 0 ? "(" : "") << p.descriptor[i] << (i < 1979 ? ", " : ")");
     return (os);
   }
+
+  std::ostream&
+  operator << (std::ostream& os, const UniqueShapeContext1960& p)
+  {
+    for (int i = 0; i < 9; ++i)
+    os << (i == 0 ? "(" : "") << p.rf[i] << (i < 8 ? ", " : ")");
+    for (std::size_t i = 0; i < 1960; ++i)
+      os << (i == 0 ? "(" : "") << p.descriptor[i] << (i < 1959 ? ", " : ")");
+    return (os);
+  }
 } // namespace pcl
